requirements.txt
# --- The Core Framework ---
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# --- Database (MongoDB) ---
motor>=3.3.0
pydantic[email]>=2.0.0

# --- Authentication (JWT & Hashing) ---
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# --- Machine Learning & Data ---
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pip install -r requirements.txt
# --- Utilities ---
python-dotenv>=1.0.0